% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areal_coverage.R
\name{calculate_areal_coverage}
\alias{calculate_areal_coverage}
\title{Calculate areal coverage}
\usage{
calculate_areal_coverage(
  exp,
  run,
  ini_date,
  ciclos,
  acumulado = 1,
  q = c(1, 5, 10, 25, 50),
  lon_band = c(-66.5, -61.5),
  lat_band = c(-35.5, -29),
  wrf_path = "/home/paola.corrales/datosmunin3/EXP/",
  out_path = "/home/paola.corrales/datosmunin3/EXP/derived_data/ppacum/"
)
}
\arguments{
\item{exp}{name of the experiment}

\item{run}{if "ana" or "fcst}

\item{ini_date}{init date and time}

\item{ciclos}{number of hourly cycles/fcst}

\item{acumulado}{time of accumulated}

\item{q}{pp threshold}

\item{lon_band}{longitud limits}

\item{lat_band}{latitude limits}

\item{wrf_path}{where the simulations are located}

\item{out_path}{where the output file will be saved}
}
\description{
Calculate areal coverage
}
