% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fss.R
\name{calculate_fss}
\alias{calculate_fss}
\title{Wrap para FSS}
\usage{
calculate_fss(fcst, obs, q, w, binary = TRUE)
}
\arguments{
\item{fcst}{matriz con precipitación pronosticada}

\item{obs}{matriz con precipitación observada (mismas dimensiones que fcst)}

\item{q}{número o vector numérico con umbrales a ser calculados}

\item{w}{número o vector numérico con escalas, correspondiente al ancho de la caja}

\item{binary}{bolean, TRUE = ambos campos son binarios.}
}
\description{
Usa la función fss del paquete {verification} pero previamente requiere que las
variables estén en matrices. También puede iterar para distintos q (valor de pp) y
w (tamaño de la caja = w2+1)
}
