% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_FSSvector_from_nonbinary_fields.R
\name{calculate_FSSvector_from_nonbinary_fields}
\alias{calculate_FSSvector_from_nonbinary_fields}
\title{Calculate the FSS for non-binary fields}
\usage{
calculate_FSSvector_from_nonbinary_fields(fbin1, fbin2, nvector)
}
\arguments{
\item{nvector}{A numeric vector containing neighborhood sizes for which the
FSS values are to be calculated. Only positive odd values are allowed in the vector.
A square neighborhood shape is assumed and the specified value represents the length
of square side.}

\item{fbind1}{A numeric matrix representing the first field.
Any values are allowed in the matrix.}

\item{fbind2}{A numeric matrix representing the first field.
Any values are allowed in the matrix. The matrix needs to have the same dimensions as fbin1.}
}
\description{
Adapted from the SpatialVx. Calculates the value of Fraction Skill Score (FSS)
for multiple neighborhood sizes.
}
