% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_matching.R
\name{calculate_probability_marching}
\alias{calculate_probability_marching}
\title{Calculate probability matching}
\usage{
calculate_probability_marching(
  exp,
  run,
  ini_date,
  ciclos,
  acumulado = 1,
  wrf_path = "/home/paola.corrales/datosmunin3/EXP/",
  out_path = "/home/paola.corrales/datosmunin3/EXP/derived_data/ppacum/"
)
}
\arguments{
\item{exp}{name of the experiment}

\item{run}{if "ana" or "fcst}

\item{ini_date}{init date and time}

\item{ciclos}{number of hourly cycles/fcst}

\item{acumulado}{time of accumulated}

\item{wrf_path}{where the simulations are located}

\item{out_path}{where the output file will be saved}
}
\description{
Calculate probability matching
}
