% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_round.R
\name{cut_round}
\alias{cut_round}
\title{Discretiza una variable en cajas y devuelve labes lindos}
\usage{
cut_round(x, breaks)
}
\arguments{
\item{x}{variable a discretizar}

\item{breaks}{vector con breaks}
}
\description{
Discretiza una variable en cajas y devuelve labes lindos
}
