% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RCRV.R
\name{get_RCRV}
\alias{get_RCRV}
\title{Calculate Reduced Centered Random Variable (RCRV)}
\usage{
get_RCRV(dt, tipo.var = "superficie", calculo = "temporal")
}
\arguments{
\item{dt}{data.table with the diag file}

\item{tipo.var}{character, superficie == variables de superficie, perfil = variables en altura,
radianzas = randianzas.}

\item{calculo}{character, temporal == calculo en el tiempo, box = por cajas, nada = por variable}
}
\description{
Calculate Reduced Centered Random Variable (RCRV)
}
