% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goes_functions.R
\name{goes_projection}
\alias{goes_projection}
\title{Goes Proyection}
\usage{
goes_projection(x, y, ncfile)
}
\arguments{
\item{x}{regular grid}

\item{y}{regular grid}

\item{ncfile}{NetCDF file}
}
\description{
Goes Proyection
}
