% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_obs_error.R
\name{input_obs_error}
\alias{input_obs_error}
\title{Input conventional observation errors from errtable to diag files}
\usage{
input_obs_error(variable, type, level, path_to_errtable = "errtable.csv")
}
\arguments{
\item{variable}{character}

\item{type}{numeric, BUFR code}

\item{level}{numeric}

\item{path_to_errtable}{path to errtable file}
}
\description{
Input conventional observation errors from errtable to diag files
}
