% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goes_functions.R
\name{rad_to_tb}
\alias{rad_to_tb}
\title{Radiance to BT}
\usage{
rad_to_tb(rad, ncfile, h = 6.629e-34, c = 299800000, kb = 1.381e-23)
}
\arguments{
\item{rad}{radiance}

\item{ncfile}{NetCDF file}

\item{h}{constant}

\item{c}{constant}

\item{kb}{constant}
}
\description{
Radiance to BT
}
