% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_diag_conv.R
\name{read_diag_conv}
\alias{read_diag_conv}
\title{Lectura de archivos diag para observaciones convencionales}
\usage{
read_diag_conv(
  file_list,
  exp,
  member = "000",
  variable = c("uv", "p", "t", "q")
)
}
\arguments{
\item{file_list}{lista de archivos}

\item{exp}{nombre del experimento asociado}

\item{member}{miembro del ensamble en caracteres. Si es la media correspode "000"}

\item{variable}{vector con las variables que se quieren obtener.}
}
\description{
Internamente define la fecha.
El formato de nombre de archivo debería ser "asim_conv_{date}.{mem}"
}
