% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_diag_rad.R
\name{read_diag_rad}
\alias{read_diag_rad}
\title{Lectura de archivos diag para radianzas}
\usage{
read_diag_rad(file_list, exp)
}
\arguments{
\item{file_list}{lista de archivos}

\item{exp}{nombre del experimento asociado}
}
\description{
Internamente define el sensor, plataforma, la fecha y el miembro del ensamble asociado.
El formato de nombre de archivo debería ser "asim_{sensor}\emph{{plat}}{date}.{mem}"
}
