% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_satbias.R
\name{read_satbias}
\alias{read_satbias}
\title{Lee archivo "satbias" de coeficientes}
\usage{
read_satbias(file)
}
\arguments{
\item{file}{archivo a leer}
}
\description{
Esta función parsea el archivo satbias y lo ordena en formato tidy.
}
