% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh.R
\name{rh}
\alias{rh}
\title{Humedad relativa}
\usage{
rh(QVAPOR, P, T)
}
\arguments{
\item{QVAPOR}{vector con la humedad especifica en kg/kg}

\item{P}{vector con la presión en Pa}

\item{T}{vector con la temperatura en kelvin}
}
\description{
Humedad relativa
}
