% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_dbz.R
\name{scale_color_dbz}
\alias{scale_color_dbz}
\title{Reflectivity color pallete}
\usage{
scale_color_dbz(
  colours = hex,
  limits = c(0, 60),
  guide = guide_colorbar(barheight = 15),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Reflectivity color pallete
}
