% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goes_functions.R
\name{scale_color_topes}
\alias{scale_color_topes}
\title{Color scale for cloud tops}
\usage{
scale_color_topes(
  colours = hex,
  limits = c(-90, 50),
  breaks = seq(-90, 50, 20),
  guide = guide_colorbar(barheight = 15),
  ...
)
}
\arguments{
\item{colours}{hex format}

\item{limits}{vector with limits}

\item{breaks}{vector with breaks}

\item{...}{others}
}
\description{
Color scale for cloud tops
}
