% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlog_trans.R
\name{symlog_trans}
\alias{symlog_trans}
\title{Transformación de escalas a symlog}
\usage{
symlog_trans(base = 10, thr = 1, scale = 1)
}
\description{
Permite transformar una escala (colores, ejes, etc.) a symlog.
}
\references{
Un buen samaritano en https://stackoverflow.com/a/14674703
}
