% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk.R
\name{tk}
\alias{tk}
\title{Temperatura}
\usage{
tk(T, P, T_BASE = 300)
}
\arguments{
\item{T}{vector con la temperatura potencial en kelvin}

\item{P}{vector con la presión en Pa}

\item{T_BASE}{valor numérico con la temperatura de base del WRF, usualmente 300 K}
}
\description{
Temperatura
}
