% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvmet.R
\name{uvmet}
\alias{uvmet}
\title{Rota los vientos e interpola a la retícula de masa}
\usage{
uvmet(ncfile)
}
\arguments{
\item{ncfile}{path al archivo netCDF donde están los datos}
}
\description{
La función calcula la rotación correcta e interpola a la reticula de masa.
Solo funciona para Labert
}
