% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_project.R
\name{wrf_project}
\alias{wrf_project}
\title{Proyecci칩n de coordenadas}
\usage{
wrf_project(
  lon,
  lat,
  inverse = FALSE,
  round = TRUE,
  map_proj = norargentina_lambert
)
}
\arguments{
\item{lon, lat}{longitud y latitud o x e y para calcular la inversa}

\item{inverse}{si es FALSE la funci칩n devuelve las coordenadas cartogr치ficas (distancia al centro del dominio).
Si es TRUE devuelve lon, lat.}

\item{round}{igual a TRUE resuelve problemas de redondeo}

\item{map_proj}{recibe un string con la proyecci칩n correspondiente en el formato aceptado por proj4}
}
\description{
Proyecta y reproyecta coordenadas utilizando el paquete proj4
}
